moved config_manager into root folder
* moved src/client/config_manager into src/config
* fixed namespace issues
*removed log_short_lived_process_command
