+++ src/tracer/src/config/bashrc_intercept.rs
