+++ src/tracer/src/lib.rs
@@ -7,13 +7,13 @@ pub mod client;
 pub mod cloud_providers;
 pub mod commands;
 pub mod common;
+pub mod config;
 pub mod daemon;
 pub mod extracts;
 pub mod logging;
 pub mod nondaemon_commands;
 pub mod process_command;
 pub mod utils;
-
 /// Runs tracer init in interactive mode
 pub fn init_command_interactive_mode(cli_args: TracerCliInitArgs) -> FinalizedInitArgs {
     InteractiveInitArgs::from_partial(cli_args)
