+++ src/tracer/src/utils.rs
@@ -1,4 +1,4 @@
-use crate::client::config_manager::TRACER_ANALYTICS_ENDPOINT;
+use crate::config::TRACER_ANALYTICS_ENDPOINT;
 use crate::common::types::analytics::{AnalyticsEventType, AnalyticsPayload};
 use anyhow::Context;
 use reqwest::Client;
